+++ lib/InlineFilters.tsx
@@ -16,6 +16,7 @@ interface BaseInlineFilters {
   toggle?: FilterTogglerType;
   onReset: () => void;
   onChange: (object: any) => void;
+  resetRender?: React.ReactNode
 }
 
 interface InlineFiltersWithDefaultValue extends BaseInlineFilters {
@@ -40,6 +41,7 @@ const InlineFilters: React.FC<
     resetText,
     toggle,
     onReset,
+    resetRender
   } = props;
 
   const [hiddenFilters, setHiddenFilters] = useLocalStorageState<string[]>(
@@ -60,7 +62,6 @@ const InlineFilters: React.FC<
 
   const { run: handleChange } = useDebounceFn(
     (values) => {
-      if (debug) console.log("INLINE FILTERS / handleChange", values);
       if (props.onChange) props.onChange(values);
     },
     { wait: delay }
@@ -93,6 +94,12 @@ const InlineFilters: React.FC<
       )
     );
   };
+
+  const resetComponent = onReset && (
+    resetRender 
+      ? <span onClick={handleReset}>{resetRender}</span> 
+      : <Button type="text" onClick={handleReset}>{resetText || "Reset filters"}</Button>
+  )
 
   let fields = schema;
   if (toggle && hiddenFilters && hiddenFilters.length > 0)
@@ -132,11 +139,7 @@ const InlineFilters: React.FC<
           {...(toggle || {})}
         />
       )}
-      {onReset && (
-        <Button type="text" onClick={handleReset}>
-          {resetText || "Reset filters"}
-        </Button>
-      )}
+      {resetComponent}
     </Space>
   );
 };