+++ CHANGELOG.md
@@ -2,4 +2,117 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.0.10](https://github.com/9troisquarts/inline-filters/compare/v0.0.9...v0.0.10) (2024-04-03)
+### [0.0.18](https://github.com/9troisquarts/inline-filters/compare/v0.0.17...v0.0.18) (2024-06-18)
+
+
+### Bug Fixes
+
+* keywords filters fixes ([07503ff](https://github.com/9troisquarts/inline-filters/commit/07503ff0f50548d660ba0f5b3501f1552dcf7946))
+
+### [0.0.17](https://github.com/9troisquarts/inline-filters/compare/v0.0.16...v0.0.17) (2024-06-10)
+
+
+### Bug Fixes
+
+* delete hover style on filters keyword options ([81de5a4](https://github.com/9troisquarts/inline-filters/commit/81de5a40d4ea85450541e0f07a17a73d5dfad7d4))
+* empty option on keywords filter ([60a566f](https://github.com/9troisquarts/inline-filters/commit/60a566fa7f94932de0ee38d5b41ba3ccc12ca4d7))
+* keywords filter words highlight ([b4ae5eb](https://github.com/9troisquarts/inline-filters/commit/b4ae5eb45daa37f060b07debdc6a477c98bf7541))
+
+### [0.0.16](https://github.com/9troisquarts/inline-filters/compare/v0.0.15...v0.0.16) (2024-05-27)
+
+
+### Bug Fixes
+
+* adds focused to keywords filter popover ([b8fa1e1](https://github.com/9troisquarts/inline-filters/commit/b8fa1e1c1b3e2700ae2eeac6b08b837342c17afc))
+* change select component of keywords filter to autocomplete ([16518b1](https://github.com/9troisquarts/inline-filters/commit/16518b110a34745ce61eb71d463bbf3e2dee38fe))
+
+### [0.0.15](https://github.com/9troisquarts/inline-filters/compare/v0.0.14...v0.0.15) (2024-05-24)
+
+### [0.0.14](https://github.com/9troisquarts/inline-filters/compare/v0.0.13...v0.0.14) (2024-05-21)
+
+
+### Bug Fixes
+
+* avoid error 'Cannot read properties of undefined (reading 'map')' ([9efba8c](https://github.com/9troisquarts/inline-filters/commit/9efba8ca8a8d165bc8cb776b6e478d246443a175))
+
+### [0.0.13](https://github.com/9troisquarts/inline-filters/compare/v0.0.12...v0.0.13) (2024-05-21)
+
+
+### Bug Fixes
+
+* fix last commit ([c1845a4](https://github.com/9troisquarts/inline-filters/commit/c1845a48cd7e38893ccb68285d1b9e8c06aa118e))
+
+### [0.0.12](https://github.com/9troisquarts/inline-filters/compare/v0.0.11...v0.0.12) (2024-05-21)
+
+
+### Bug Fixes
+
+* (keywords) ensure that results are loaded ([20eefb8](https://github.com/9troisquarts/inline-filters/commit/20eefb84c8698b35741258ffcd5a157bc27d29a8))
+
+### [0.0.11](https://github.com/9troisquarts/inline-filters/compare/v0.0.10...v0.0.11) (2024-05-21)
+
+### [0.0.10](https://github.com/9troisquarts/inline-filters/compare/v0.0.9...v0.0.10) (2024-05-21)
+
+
+### Features
+
+* keywords filter ([913f789](https://github.com/9troisquarts/inline-filters/commit/913f789e631879d4ecb8afbca9dd829083d1e9c8))
+* propriété i18n ([c731035](https://github.com/9troisquarts/inline-filters/commit/c7310359f4e2e2cb4516a746af085421b3f15894))
+
+
+### Bug Fixes
+
+* retours alex/kevin sur filter keywords ([c9bd4e0](https://github.com/9troisquarts/inline-filters/commit/c9bd4e08de88a0c3d2e095e762fab49a444eb2f5))
+
+### [0.0.9](https://github.com/9troisquarts/inline-filters/compare/v0.0.8...v0.0.9) (2023-12-06)
+
+
+### Bug Fixes
+
+* hasOwn is experimental ([618d226](https://github.com/9troisquarts/inline-filters/commit/618d2267211f3b3bd12ca912200176ef77a50c7d))
+
+### [0.0.8](https://github.com/9troisquarts/inline-filters/compare/v0.0.7...v0.0.8) (2023-12-06)
+
+
+### Bug Fixes
+
+* onReset should reset the internal state when uncontrolled ([0c85ec5](https://github.com/9troisquarts/inline-filters/commit/0c85ec5c34d909df537b93207c94850771cac66e))
+
+### [0.0.7](https://github.com/9troisquarts/inline-filters/compare/v0.0.6...v0.0.7) (2023-12-06)
+
+
+### Bug Fixes
+
+* skip useeffect if value is undefined ([5e12b6b](https://github.com/9troisquarts/inline-filters/commit/5e12b6b5153df6376bb3d5ff7d119a6753d983a0))
+
+### [0.0.6](https://github.com/9troisquarts/inline-filters/compare/v0.0.5...v0.0.6) (2023-12-06)
+
+
+### Bug Fixes
+
+* Retrait du label ([bcad855](https://github.com/9troisquarts/inline-filters/commit/bcad855a0c62f50f04d21e76f0ff94274f3c38a3))
+
+### [0.0.5](https://github.com/9troisquarts/inline-filters/compare/v0.0.4...v0.0.5) (2023-12-06)
+
+
+### Features
+
+* disable commitlint ([2db5c35](https://github.com/9troisquarts/inline-filters/commit/2db5c35ec0a7ff7ddabc73e6577cc2410e624dfb))
+
+### 0.0.4 (2023-12-06)
+
+### 0.0.3 (2023-12-06)
+
+### 0.0.2 (2023-12-06)
+
+### 0.0.6 (2023-12-06)
+
+### 0.0.5 (2023-12-06)
+
+### 0.0.4 (2023-12-06)
+
+### 0.0.4 (2023-12-06)
+
+### 0.0.3 (2023-12-06)
+
+### 0.0.2 (2023-12-06)